+++ nodemgr/internal/lib/nfdapi/swagger/model_offer_request_body.go
@@ -11,7 +11,7 @@ package swagger
 
 type OfferRequestBody struct {
 	// Amount in microAlgo to sell the NFD for.  1 ALGO is 1,000,000 microAlgo. 1 ALGO minimum, 100 million maximum.
-	Offer int32 `json:"offer"`
+	Offer int64 `json:"offer"`
 	// Whether to pay the offer amount + fees to the account the NFD is reserved for.  Equivalent to 'gifting' the NFD
 	PayReceiver bool   `json:"payReceiver,omitempty"`
 	ReservedFor string `json:"reservedFor,omitempty"`
