+++ nodemgr/internal/lib/nfdapi/swagger/model_post_offer_to_owner_request_body.go
@@ -13,6 +13,6 @@ type PostOfferToOwnerRequestBody struct {
 	// Note to pass along to the NFD owner.  Must be provided but can be blank
 	Note string `json:"note"`
 	// Amount in microAlgo being offered to the NFD owner
-	Offer  int32  `json:"offer"`
+	Offer  int64  `json:"offer"`
 	Sender string `json:"sender"`
 }
